requirements.txt
flask==2.2.2
werkzeug>=2.2.2
waitress==2.1.2
redis==4.3.4
psycopg2-binary==2.9.4
